Crate wasm_transform

Modules

artifact

This module defines the notion of the artifact. This is a processed and instantiated module that can have its exposed methods invoked.

constants

Core constants used in parsing and validation.

machine

An implementation of the abstract machine that can run artifacts.

metering_transformation

A program transformation that inserts metering instructions into a Wasm module.

output

Functionality for outputting Wasm modules in binary format.

parse

This module defines a parser for the Web assembly binary format conforming to the specification in wasm-core-1-20191205 but with further restrictions to ensure suitability for the Concordium blockchain.

types

AST definition of Wasm modules, as well as supporting datatypes. Based on the W3C Wasm specification

utils

Common utilities for Wasm transformations. These are wrappers around the basic functionality exposed by other modules.

validate

Utilities for Wasm module validation.