Module wasm_transform::machine

An implementation of the abstract machine that can run artifacts.

Structs

RunConfig

Configuration that can be run.

RuntimeStack

A runtime stack. This contains both the stack in a function, as well as all the function parameters and locals of the function.

Enums

ExecutionOutcome

NoInterrupt

An empty type used when no interrupt is possible by a host function call.

RuntimeError

Value

A Wasm typed value. The values are not inherently signed or unsigned, but we choose signed integers as the representation type.

Traits

Host

The host that can process external functions.

Type Definitions

RunResult

Result of execution. Runtime exceptions are returned as Err(_). This includes traps, illegal memory accesses, etc.