Trait wasm_transform::machine::Host

pub trait Host<I> {
    type Interrupt;
    fn tick_initial_memory(&mut self, num_pages: u32) -> RunResult<()>;

    fn call(
        &mut self, 
        f: &I, 
        memory: &mut Vec<u8>, 
        stack: &mut RuntimeStack
    ) -> RunResult<Option<Self::Interrupt>>;
}

The host that can process external functions.

Associated Types

type Interrupt

Required methods

fn tick_initial_memory(&mut self, num_pages: u32) -> RunResult<()>

Charge the given amount of energy for the initial memory. The argument is the number of pages.

fn call(
&mut self,
f: &I,
memory: &mut Vec<u8>,
stack: &mut RuntimeStack
) -> RunResult<Option<Self::Interrupt>>

Call the specified host function, giving it access to the current memory and stack. The return value of Ok(()) signifies that execution succeeded and the machine should proceeed, the return value of Err(_) signifies a trap.