Trait wasm_transform::parse::GetParseable

pub trait GetParseable<A, Ctx> {
    fn next(self, ctx: Ctx) -> ParseResult<A>;
}

A helper trait for more convenient use. The difference from the above is that typically the result type is determined by the context, which we take advantage of to reduce the need for typing annotations which would be needed by the Parseable trait.

The reason for that is that this trait defines a new method on the type, giving us access to all of the convenience features of Rust that come with it.

Required methods

fn next(self, ctx: Ctx) -> ParseResult<A>

Parse an item. Analogous to ‘parse’, but with the reversed roles for types of input and output. In the ‘Parseable’ trait the trait is defined for the type that is to be parsed and the source is fixed, whereas here the trait is parameterized by the type to be parsed, and the trait is implemented for the source type.