Trait wasm_transform::validate::HasValidationContext

pub trait HasValidationContext {
    fn get_local(&self, idx: LocalIndex) -> ValidateResult<ValueType>;

    fn get_global(&self, idx: GlobalIndex) -> ValidateResult<(ValueType, bool)>;

    fn memory_exists(&self) -> bool;

    fn table_exists(&self) -> bool;

    fn get_func(&self, idx: FuncIndex) -> ValidateResult<&Rc<FunctionType>>;

    fn get_type(&self, idx: TypeIndex) -> ValidateResult<&Rc<FunctionType>>;

    fn return_type(&self) -> BlockType;
}

The trait used used to parametrize the validation algorithm so that it can be used for other applications than mere validation. In particular the validation algorithm maintains useful state during its run, e.g., current and maximum stack height, which is useful during compilation.

Required methods

fn get_local(&self, idx: LocalIndex) -> ValidateResult<ValueType>

Get the local of a function at the given index. Note that function parameters define implicit locals.

fn get_global(&self, idx: GlobalIndex) -> ValidateResult<(ValueType, bool)>

Get a global together with its mutability. true for mutable, false for constant.

fn memory_exists(&self) -> bool

Return whether the module has memory.

fn table_exists(&self) -> bool

Return whether the module has the table.

fn get_func(&self, idx: FuncIndex) -> ValidateResult<&Rc<FunctionType>>

Get the type of the function at the given index.

fn get_type(&self, idx: TypeIndex) -> ValidateResult<&Rc<FunctionType>>

Get the type at the given index.

fn return_type(&self) -> BlockType

Return the return type of the function.